TW-87415 terraform E-Mail settings: fix optional fields
Also, TC rest api returns empty string when login is not provided and null for password if it is not provided. This caused the plan to have stateValueNull both for login and password when importing but after applying login state value was not null and this was reported as a bug in provider from terraform side. Now login value defaults to empty string in the state.